+++ packages/docs/src/components/ColorsGrid.tsx
@@ -1,4 +1,4 @@
-import { colors } from '@joaogomesdev-ignite-ui/tokens'
+import { colors } from '@joaogomesdev-studies/tokens'
 import { getContrast } from 'polished'
 
 export function ColorsGrid() {