+++ packages/docs/src/stories/Box.stories.tsx
@@ -1,5 +1,5 @@
 import type { StoryObj, Meta } from '@storybook/react'
-import { Box, BoxProps, Text } from '@joaogomesdev-ignite-ui/react'
+import { Box, BoxProps, Text } from '@joaogomesdev-studies/react'
 
 export default {
   title: 'Surfaces/Box',