+++ packages/docs/src/stories/Checkbox.stories.tsx
@@ -1,10 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import {
-  Box,
-  Text,
-  Checkbox,
-  CheckboxProps,
-} from '@joaogomesdev-ignite-ui/react'
+import { Box, Text, Checkbox, CheckboxProps } from '@joaogomesdev-studies/react'
 
 export default {
   title: 'Form/Checkbox',