+++ packages/docs/src/stories/Text.stories.tsx
@@ -1,5 +1,5 @@
 import type { StoryObj, Meta } from '@storybook/react'
-import { TextProps, Text } from '@joaogomesdev-ignite-ui/react'
+import { TextProps, Text } from '@joaogomesdev-studies/react'
 
 export default {
   title: 'Typography/Text',